# speed up `format_classprobs()` helper